requirements.txt
appdirs==1.4.3
Django==1.11.1
django-simple-history==1.8.2
django-tables2==1.7.1
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
virtualenv==15.1.0
